requirements.txt
aiogram==3.7.0
SQLAlchemy>=2.0
pandas
openpyxl
python-dotenv
